requirements.txt
websockets>=15.0.1
mss>=10.1.0
Pillow>=12.0.0
pyobjc>=12.0.0
